+++ packages/js-utils/src/index.ts
@@ -5,7 +5,6 @@ import { Changelogs } from './Changelogs'
 import { Collection } from './Collection'
 import { Events } from './Events'
 import { Functions } from './Functions'
-import { Hashes } from './Hashes'
 import { Ids } from './Ids'
 import { Numbers } from './Numbers'
 import { Objects } from './Objects'
@@ -22,7 +21,6 @@ export {
   Combinations,
   Events,
   Functions,
-  Hashes,
   Ids,
   nGram,
   Numbers,
@@ -33,8 +31,5 @@ export {
   Units
 }
 
-
 export * from './types'
-export type {
-  CommonLogger
-}
+export type { CommonLogger }