+++ packages/node-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.5.20
 
+## 0.8.5
+
+### Patch Changes
+
+- Secret and Security
+
 ## 0.8.4
 
 ### Patch Changes